[frontend] ZATSU full screen pong board
